Sixten snart färdig

+++ Assets/Scripts/Player/PlayerHealth.cs
@@ -6,7 +6,7 @@ public class PlayerHealth : MonoBehaviour
     private float playerHealth;
     [SerializeField] PlayerData _playerData;
 
-    public void PlayerTakeDamage(float damage)
+    public void PlayerTakeDamage(int damage) // Ändrade från float till int
     {
         _playerData.currentHealth -= damage;
         // Debug.Log($"Player health {_playerData.currentHealth}");
